requirements.txt
flask
rich
jinja2
requests
python-dotenv